+++ osu.Framework/Resources/Shaders/sh_CircularProgress.fs
@@ -7,9 +7,9 @@
 
 varying highp vec2 v_TexCoord;
 
-uniform lowp sampler2D m_Sampler;
-uniform mediump float progress;
-uniform mediump float innerRadius;
+uniform highp sampler2D m_Sampler;
+uniform highp float progress;
+uniform highp float innerRadius;
 uniform highp float texelSize;
 uniform bool roundedCaps;
 
@@ -19,7 +19,7 @@ void main(void)
     highp vec2 pixelPos = v_TexCoord / resolution;
     
     highp vec2 wrappedCoord = wrap(v_TexCoord, v_TexRect);
-    lowp vec4 textureColour = getRoundedColor(wrappedSampler(wrappedCoord, v_TexRect, m_Sampler, -0.9), wrappedCoord);
+    highp vec4 textureColour = getRoundedColor(wrappedSampler(wrappedCoord, v_TexRect, m_Sampler, -0.9), wrappedCoord);
 
     gl_FragColor = vec4(textureColour.rgb, textureColour.a * progressAlphaAt(pixelPos, progress, innerRadius, roundedCaps, texelSize));
 }
